# slackbot-openapi-spec.json
{
  "info": {
    "title": "Events API for Watson Assistant",
    "version": "0.8.1"
  },
  "tags": [],
  "servers": [
    {
      "description": "Code Engine deployment",
      "url": "https://{appname}.{projectid}.{region}.codeengine.appdomain.cloud",
      "variables": {
        "appname": {
          "default": "myapp",
          "description": "application name"
        },
        "projectid": {
          "default": "projectid",
          "description": "the Code Engine project ID"
        },
        "region": {
          "default": "us-south",
          "description": "the deployment region, e.g., us-south"
        }
      }
    },
    {
      "description": "local test",
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "5000",
          "description": "local port to use"
        }
      }
    }
  ],
  "paths": {
    "/": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Greeting",
        "description": "health check"
      }
    },
    "/events": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 30
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsOut"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "all events",
        "description": "Retrieve all event records",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "parameters": [],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "Insert a new event record",
        "description": "Insert a new event record with the given attributes. Its new EID is returned.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/database/recreate": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "confirmation",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "Recreate the database schema",
        "description": "Recreate the database schema and insert sample data.\nRequest must be confirmed by passing query parameter.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/events/eid/{eid}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Event record by EID",
        "description": "Retrieve a single event record by its EID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Delete an event record by EID",
        "description": "Delete a single event record identified by its EID.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/events/name/{short_name}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "short_name",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Event record by name",
        "description": "Retrieve a single event record by its short name",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "openapi": "3.0.3",
  "components": {
    "schemas": {
      "ValidationError": {
        "properties": {
          "detail": {
            "type": "object",
            "properties": {
              "<location>": {
                "type": "object",
                "properties": {
                  "<field_name>": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "EventOut": {
        "type": "object",
        "properties": {
          "shortname": {
            "type": "string"
          },
          "enddate": {
            "type": "string",
            "format": "date"
          },
          "begindate": {
            "type": "string",
            "format": "date"
          },
          "eid": {
            "type": "integer"
          },
          "contact": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "per_page": {
            "type": "integer"
          },
          "current": {
            "type": "string",
            "format": "url"
          },
          "prev": {
            "type": "string",
            "format": "url"
          },
          "pages": {
            "type": "integer"
          },
          "next": {
            "type": "string",
            "format": "url"
          },
          "last": {
            "type": "string",
            "format": "url"
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "first": {
            "type": "string",
            "format": "url"
          }
        }
      },
      "EventsOut": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventOut"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "EventIn": {
        "type": "object",
        "properties": {
          "shortname": {
            "type": "string",
            "minLength": 0,
            "maxLength": 20
          },
          "enddate": {
            "type": "string",
            "format": "date"
          },
          "begindate": {
            "type": "string",
            "format": "date"
          },
          "contact": {
            "type": "string",
            "minLength": 0,
            "maxLength": 255
          },
          "location": {
            "type": "string",
            "minLength": 0,
            "maxLength": 60
          }
        },
        "required": [
          "begindate",
          "contact",
          "enddate",
          "location",
          "shortname"
        ]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}